updated readme and milestoneBar WIP
