+++ ui/src/routes/ProjectStatus.js
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import { useParams, useNavigate } from 'react-router-dom';
 import { Button, Paper, Typography, Box, Divider, TextField, List, ListItem, Avatar } from '@mui/material';
 import { useCookies } from 'react-cookie';
+import MilestoneBar from '../components/milestoneBar';
 import './ProjectStatus.css';
 
 
@@ -17,110 +18,6 @@ const ProjectStatus = () => {
     const [chatposts, setChatposts] = useState([]);
     const [userdata, setUserdata] = useState([])
     const [currentUserDoubloons, setCurrentUserDoubloons] = useState()
-
-    const milestones = [
-        "Kickoff",
-        "Developing",
-        "Testing",
-        "Staging",
-        "Seeking Funds",
-        "Deploy",
-        "Sustainment"
-    ];
-
-    const [currentMilestoneIndex, setCurrentMilestoneIndex] = useState(-1);
-    const [milestoneTimestamps, setMilestoneTimestamps] = useState(
-        milestones.map(() => ({ started: null, completed: null }))
-    );
-
-    const [canStartCurrentMilestone, setCanStartCurrentMilestone] = useState(true);
-    const [canCompleteCurrentMilestone, setCanCompleteCurrentMilestone] = useState(false);
-
-    const getStatusClass = (index) => {
-        if (index === currentMilestoneIndex) {
-            return 'active'; 
-        } else if (index < currentMilestoneIndex) {
-            return 'completed'; 
-        }
-        return 'inactive'; 
-    };
-
-    const moveToPreviousMilestone = () => {
-        if (currentMilestoneIndex > 0) {
-            setCurrentMilestoneIndex(currentMilestoneIndex - 1);
-
-            updateCompleteButtonState(currentMilestoneIndex - 1);
-        }
-    };
-
-    const moveToNextMilestone = () => {
-        if (currentMilestoneIndex < milestones.length - 1) {
-            setCurrentMilestoneIndex(currentMilestoneIndex + 1);
-
-            updateCompleteButtonState(currentMilestoneIndex + 1);
-            setCanStartCurrentMilestone(true);
-            setCanCompleteCurrentMilestone(true);
-        }
-    };
-
-    const startCurrentMilestone = () => {
-        if (canStartCurrentMilestone && currentMilestoneIndex >= 0) {
-            const now = new Date().toLocaleString();
-            setMilestoneTimestamps(prev => {
-                const updated = [...prev];
-                updated[currentMilestoneIndex].started = now;
-                return updated;
-            });
-            setCanStartCurrentMilestone(false);
-            setCanCompleteCurrentMilestone(true);
-        }
-    };
-
-    const completeCurrentMilestone = () => {
-        if (canCompleteCurrentMilestone && currentMilestoneIndex >= 0 && milestoneTimestamps[currentMilestoneIndex].started) {
-            const now = new Date().toLocaleString();
-            setMilestoneTimestamps(prev => {
-                const updated = [...prev];
-                updated[currentMilestoneIndex].completed = now;
-                return updated;
-            });
-            setCanCompleteCurrentMilestone(false);
-        }
-    };
-
-    const removeTimestamp = () => {
-        if (currentMilestoneIndex >= 0 && milestoneTimestamps[currentMilestoneIndex]) {
-            setMilestoneTimestamps(prev => {
-                const updated = [...prev];
-                const currentMilestone = updated[currentMilestoneIndex];
-
-
-                if (currentMilestone.completed) {
-                    currentMilestone.completed = null;
-                    setCanCompleteCurrentMilestone(true);
-                } else if (currentMilestone.started) {
-                    currentMilestone.started = null;
-                    setCanStartCurrentMilestone(true);
-                }
-
-                return updated;
-            });
-        }
-    };
-    
-
-
-    const updateCompleteButtonState = (index) => {
-        if (milestoneTimestamps[index]?.started && !milestoneTimestamps[index]?.completed) {
-            setCanCompleteCurrentMilestone(true);
-        } else {
-            setCanCompleteCurrentMilestone(false);
-        }
-    };
-
-
-
-
 
     const handleAddComment = () => {
         if (newComment.trim()) {
@@ -438,96 +335,10 @@ const ProjectStatus = () => {
                         style={{ fontWeight: "500", color: "#616161" }}>
                         <p className='product-status'>Application Development Status:</p>
                         <div>
-                            <div className="milestone-wrapper">
-                                <div className="milestone-container">
-                                    {milestones.map((milestone, index) => (
-                                        <div key={index} className={`milestone ${getStatusClass(index)}`}>
-                                            {milestone}
-                                        </div>
-                                    ))}
-                                </div>
-
-                                <div className="button-container">
-                                    <button
-                                        className="milestone-button"
-                                        onClick={moveToPreviousMilestone}
-                                        disabled={currentMilestoneIndex <= 0} // Disable when at the first milestone
-                                    >
-                                        Previous Milestone
-                                    </button>
-
-                                    <button
-                                        className="milestone-button"
-                                        onClick={moveToNextMilestone}
-                                        disabled={currentMilestoneIndex >= milestones.length - 1} // Disable when at the last milestone
-                                    >
-                                        Next Milestone
-                                    </button>
-
-                                    <button
-                                        className="milestone-button"
-                                        onClick={removeTimestamp}
-                                        disabled={milestoneTimestamps[currentMilestoneIndex]?.started === null && milestoneTimestamps[currentMilestoneIndex]?.completed === null}
-                                    >
-                                        Remove Timestamp
-                                    </button>
-
-                                    <button
-                                        className="milestone-button"
-                                        onClick={startCurrentMilestone}
-                                        disabled={!canStartCurrentMilestone || currentMilestoneIndex < 0}
-                                    >
-                                        Start Milestone
-                                    </button>
-
-                                    <button
-                                        className="milestone-button"
-                                        onClick={completeCurrentMilestone}
-                                        disabled={!(milestoneTimestamps[currentMilestoneIndex]?.started) || currentMilestoneIndex < 0}
-                                    >
-                                        Complete Milestone
-                                    </button>
-                                </div>
-
-                                <div className="timestamp-list">
-                                    <h3>Milestone History</h3>
-                                    <ul>
-                                        {milestones.map((milestone, index) => (
-                                            <li key={index}>
-                                                <strong>{milestone}</strong>
-                                                <div>
-                                                    {milestoneTimestamps[index]?.started && (
-                                                        <small>Started on: {milestoneTimestamps[index].started}</small>
-                                                    )}
-                                                </div>
-                                                <div>
-                                                    {milestoneTimestamps[index]?.completed && (
-                                                        <small className='completed-stamp'>Completed on: {milestoneTimestamps[index].completed}</small>
-                                                    )}
-                                                </div>
-                                            </li>
-                                        ))}
-                                    </ul>
-                                </div>
-                            </div>
+                            <MilestoneBar />
                         </div>
 
                     </Typography>
-                    {/* <Typography
-                        paragraph
-                        style={{
-                            fontSize: "1rem",
-                            marginTop: "0.5rem",
-                            marginBottom: "1.5rem",
-                        }}>
-                        {bounty.problem_statement}
-                    </Typography>
-                    <Typography
-                        variant="h6"
-                        style={{ fontWeight: "500", color: "#616161" }}>
-                        Submitter ID: {bounty.submitter_id}
-                    </Typography> */}
-                    {/* Git Text render */}
 
                     <Typography
                         variant="h6"
